backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
email-validator==2.0.0.post2
fastapi==0.115.6
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
spotipy==2.23.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0